+++ test/td_cx/jobs/job_test.exs
@@ -1,4 +1,4 @@
-defmodule TdCx.Sources.Jobs.JobTest do
+defmodule TdCx.Jobs.JobTest do
   use TdCx.DataCase
 
   alias Elasticsearch.Document
@@ -6,7 +6,8 @@ defmodule TdCx.Sources.Jobs.JobTest do
 
   describe "Job" do
     test "encode/1 without events" do
-      %{external_id: external_id, id: id, source: source} = job = insert(:job, events: [])
+      %{external_id: external_id, id: id, source: source} =
+        job = insert(:job, events: [], type: "foo")
 
       assert Document.encode(job) == %{
                external_id: external_id,
@@ -15,7 +16,8 @@ defmodule TdCx.Sources.Jobs.JobTest do
                  external_id: source.external_id,
                  type: source.type
                },
-               status: ""
+               status: "",
+               type: "foo"
              }
     end