+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [TD-2949] Support for creating connector jobs in Kubernetes
 - [TD-3089] Widget and type `copy` on df
 
 ## [4.8.0] 2020-11-16